Authorization

+++ my-mean/backend/routes/posts.js
@@ -67,10 +67,13 @@ router.put(
     content: req.body.content,
     imagePath: imagePath
   })
-  console.log(post)
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    res.status(200).json({message: "Update Successful!"})
-  })
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({message: "Update Successful!"})
+    } else {
+      res.status(401).json({message: "Not Authorized"})
+    }
+  });
 });
 
 router.get("", (req, res, next) => {
@@ -108,9 +111,13 @@ router.get("/:id", (req, res, next) => {
 })
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result)
-    res.status(200).json({ message: "Post deleted!" });
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+    console.log(result);
+    if (result.n > 0) {
+      res.status(200).json({message: "Deletion Successful!"})
+    } else {
+      res.status(401).json({message: "Not Authorized"})
+    }
   })
 });
